feat(version4): FIMD Global (multi-output + work-hours) con muestras FORZADAS en torno a pulsos de N - Qué agrega respecto al GLOBAL anterior: 1) Para CADA fermentación, si un pulso de N cae en horario hábil (L–V, 08–17), se fuerza: (a) **una muestra en el momento del pulso** (redondeada al slot de muestreo más cercano permitido) y (b) **≥1 muestra dentro de [tN, tN+2 h]**. Ambas cuentan para el tope de **2 muestras/día/exp**. 2) Si forzar ambas excede el tope diario, se prioriza la muestra **en el pulso** y se intenta reubicar la post-pulso en el siguiente día hábil cercano. 3) Selección GLOBAL greedy (D-óptimo rank-m) con presupuesto K_total repartido entre todas las fermentaciones elegidas. 4) Exporta un plan en Excel con resumen de diseños y calendario de muestreo.
Notas:
- tauN (τ) es la **duración (h)** del pulso rectangular de N. El modelo reparte la dosis (mg/L) en un caudal constante durante τ horas (term source en dN/dt). Físicamente aproxima una adición con mezcla distribuida en τ (evita delta-impulso)
